requirements.txt
gymnasium[mujoco]
h5py
matplotlib
pyyaml
tensorboard
torch
tqdm
yattag